packs: device: create flash regions for algorithm elements with no memory element
This change addresses a long-standing issue by creating flash memory
regions for pack based targets when there is an <algorithm> element
that does not have a corresponding <memory> element. Only algorithms
marked as default are converted in this way. Also, the region will
be marked as non-testable.

Note that any cases where there are multiple default algorithms that
have overlapping address ranges will result in an exception. This
should be acceptable because there aren't any such cases in the wild
to my knowledge.